refactor: 상수 도메인별 파일로 분리
