+++ src/car.js
@@ -2,7 +2,7 @@ import {
   DEFAULT_STEP_SIZE,
   MAX_NAME_LENGTH,
   MIN_NAME_LENGTH
-} from './constants/app'
+} from './constants/car'
 import { CAR_ERROR_MESSAGE } from './constants/error'
 import { isString } from './utils/validator'
