+++ src/constants/car.js
@@ -1,0 +1,4 @@
+export const DEFAULT_STEP_SIZE = 1
+export const RUN_THRESHOLDS = 4
+export const MAX_NAME_LENGTH = 5
+export const MIN_NAME_LENGTH = 1
