+++ src/constants/race.js
@@ -1,14 +1,10 @@
-export const DEFAULT_STEP_SIZE = 1
-export const RUN_THRESHOLDS = 4
-export const MAX_NAME_LENGTH = 5
-export const MIN_NAME_LENGTH = 1
-
 export const REQUIRE_METHODS_KEY = [
   'getName',
   'getPosition',
   'setPosition',
   'run'
 ]
+
 export const DEFAULT_MAX_MATCH_LENGTH = 5
 export const DEFAULT_RUN_CONDITION = () => true
 export const MIN_PARTICIPANTS_LENGTH = 2
